move dest of copy to tarcopy
